requirement.txt
pyserial
opencv-python
numpy
pymavlink